+++ include/mppi/controllers/mppi.cuh
@@ -161,8 +161,8 @@ class MPPIController {
     HANDLE_ERROR(
         cudaMalloc(&d_u_nom_, config.horizon * config.nu * sizeof(float)));
     /* curandGenerateNormal requires an even element count. */
-    int noise_count_ = config.num_samples * config.horizon * config.nu;
-    noise_count_ += (noise_count_ & 1);
+    size_t noise_count_ = static_cast<size_t>(config.num_samples) * config.horizon * config.nu;
+    noise_count_ += (noise_count_ & 1u);
     HANDLE_ERROR(cudaMalloc(&d_noise_, noise_count_ * sizeof(float)));
     HANDLE_ERROR(cudaMalloc(&d_costs_, config.num_samples * sizeof(float)));
     HANDLE_ERROR(cudaMalloc(&d_initial_state_, config.nx * sizeof(float)));
@@ -292,8 +292,8 @@ class MPPIController {
       }
 
       // Sample N(0,1) noise
-      int noise_n = config_.num_samples * config_.horizon * config_.nu;
-      noise_n += (noise_n & 1);
+      size_t noise_n = static_cast<size_t>(config_.num_samples) * config_.horizon * config_.nu;
+      noise_n += (noise_n & 1u);
       HANDLE_CURAND_ERROR(curandGenerateNormal(gen_, d_noise_, noise_n, 0.0f, 1.0f));
 
       // Apply reference bias to alpha fraction of samples (split sampling)
